requirements.txt
# Monsoon Failures & Textile Stock Volatility Prediction System
# Production Requirements

# Core Data
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Data Acquisition
yfinance>=0.2.30
imdlib>=0.1.15
requests>=2.31.0

# Econometrics & Statistical Models
arch>=6.0.0
statsmodels>=0.14.0
scipy>=1.11.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
optuna>=3.4.0

# Deep Learning (optional — install manually if needed for LSTM layer)
# tensorflow>=2.15.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.30.0

# API & LLM
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.4.0

# Utilities
openpyxl>=3.1.0
joblib>=1.3.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
python-telegram-bot>=21.0
gTTS>=2.4.0

# ── Security additions ──────────────────────────────────────────────────
# Email validation (RFC 5322 compliant)
email-validator>=2.1.0
# Password hashing for admin panel
bcrypt>=4.1.0
# python-multipart required by FastAPI for form data
python-multipart>=0.0.9
